addresses i#70. seperated out each crud operation into its own files. modified pressDataDecoder from ParseABI to some vanilla js.
