+++ apps/backend/src/decodingSchema/decoders/pressDataDecoder.ts
@@ -11,10 +11,35 @@ import { publicClient } from "../../viem/client";
 
   export type PressData = {};
 
-  export async function pressDataDecoder(press: Hex) {
-    const bytecode = await publicClient.getBytecode({
-        address: press
-    })
-    console.log(bytecode)
-
+  export function stringDecoder(data: Hash): readonly [String] {
+    return decodeAbiParameters(
+      parseAbiParameters(
+        "string"
+      ),
+      data
+    );
   }
+  
+  export async function pressDataDecoder(press: `0x${string}`) {
+    if (!press) {
+      return undefined;
+    }
+  
+    const bytecode: any = await publicClient.getBytecode({
+      address: press,
+    });
+  
+    if (!bytecode) {
+      return undefined;
+    }
+  
+    const stringData = bytecode.slice(64);
+    const decodedString = Buffer.from(stringData, 'hex').toString('utf8').replace(/\0/g, '');
+  
+    if (!decodedString) {
+      return undefined;
+    }
+  
+    const cleanedContractUri = decodedString.replace(/\u0011/g,'').trim();
+    return cleanedContractUri;
+}
